requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
python-dotenv
pydantic
httpx
python-jose
passlib[bcrypt]
psycopg2-binary
aiodns
pytest
python-multipart
pydantic-settings
pydantic[email]
greenlet
openai
langchain_groq
langchain_core